# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "DigiSchool API",
        "description": "API REST pour la gestion d'ecoles DigiSchool.\n\n## Authentification\nLa plupart des endpoints necessitent un token JWT.\n1. Utilisez `POST /api/auth/login` pour obtenir un token\n2. Cliquez sur le bouton \"Authorize\" et entrez: `Bearer <votre_token>`\n\n## Multi-tenancy\nLe systeme est multi-tenant base sur l'ecole (ecoleId).\nL'ecoleId est automatiquement extrait du token JWT.\n\n## Comptes de test\n| Role | Email | Telephone | Password |\n|------|-------|-----------|----------|\n| Admin | admin@digischool.cm | +237600000000 | Admin@2025 |\n| Directeur | smbarga@lavictoire.cm | +237677123456 | Directeur@2025 |\n| Enseignant | jpkamga@lavictoire.cm | +237670111222 | Enseignant@2025 |\n| Secretaire | catangana@lavictoire.cm | +237660555666 | Secretaire@2025 |\n| Parent | fnkoulou@gmail.com | +237691666777 | Parent@2025 |\n",
        "contact": {
            "name": "DigiSchool Support",
            "url": "https://digischool.cm",
            "email": "support@digischool.cm"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://digischool.cm/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Serveur de developpement local"
        },
        {
            "url": "https://api.digischool.cm",
...
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/users/{id}/status": {
            "put": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "updateUserStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ACTIVE",
                                "INACTIVE",
                                "LOCKED",
                                "PENDING"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "getUsers",
                "parameters": [
                    {
                        "name": "tenantId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "SUPER_ADMIN",
                                "ADMIN_ECOLE",
                                "ENSEIGNANT",
                                "PARENT",
                                "SECRETAIRE",
                                "COMPTABLE"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserInfo"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserInfo"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserInfo": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "type": "string"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "prenom": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "SUPER_ADMIN",
                            "ADMIN_ECOLE",
                            "ENSEIGNANT",
                            "PARENT",
                            "SECRETAIRE",
                            "COMPTABLE"
                        ]
                    },
                    "tenantId": {
                        "type": "string"
                    }
                }
            },
...
            "RegisterRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 2147483647,
                        "minLength": 8
                    },
                    "nom": {
                        "type": "string",
                        "minLength": 1
                    },
                    "prenom": {
                        "type": "string",
                        "minLength": 1
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "SUPER_ADMIN",
                            "ADMIN_ECOLE",
                            "ENSEIGNANT",
                            "PARENT",
                            "SECRETAIRE",
                            "COMPTABLE"
                        ]
                    },
                    "tenantId": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "role"
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtenu via POST /api/auth/login. Entrez: Bearer <token>",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}